backend/requirements.txt
flask
flask-cors
groq
papermage
transformers
torch
sentence-transformers
python-dotenv
gunicorn
